+++ src/types/recursivePartial.ts
@@ -1,0 +1,11 @@
+/**
+ * Utility type that makes partial all nested properties
+ */
+export type RecursivePartial<T> = 
+{
+    [P in keyof T]?: T[P] extends (infer U)[] ?
+                        RecursivePartial<U>[] : 
+                        T[P] extends object | undefined ?
+                            RecursivePartial<T[P]> :
+                            T[P];
+};